+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Renamed DeliveryMandate to Mandate.
 - Rename _Abgabe_ to _Lieferung_.
+- Refactored delivery overview to use only [MUI](https://mui.com/material-ui/) components.
 
 ## v1.0.93 - 2024-05-14